requirements-dev.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
--no-binary psycopg2

appdirs==1.4.4            # via virtualenv
autoflake==1.4
autopep8==1.5.4
backcall==0.2.0           # via ipython
decorator==4.4.2          # via ipython, traitlets
distlib==0.3.1            # via virtualenv
django-extensions==3.0.5
filelock==3.0.12          # via tox, virtualenv
importlib-metadata==1.7.0  # via pluggy, tox, virtualenv
importlib-resources==3.0.0  # via virtualenv
ipython==7.9.0
ipython-genutils==0.2.0   # via traitlets
isort==4.3.21
jedi==0.17.2              # via ipython
packaging==20.4           # via tox
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.13.1            # via tox
prompt-toolkit==2.0.10    # via ipython
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via tox
pycodestyle==2.6.0        # via autopep8
pydocstyle==5.1.0
pyflakes==2.2.0           # via autoflake
pygments==2.6.1           # via ipython
pyparsing==2.4.7          # via packaging
six==1.15.0               # via packaging, prompt-toolkit, tox, traitlets, virtualenv
snowballstemmer==2.0.0    # via pydocstyle
toml==0.10.1              # via autopep8, tox
tox==3.19.0
traitlets==4.3.3          # via ipython
virtualenv==20.0.31       # via tox
wcwidth==0.2.5            # via prompt-toolkit
werkzeug==1.0.1
zipp==1.2.0               # via importlib-metadata, importlib-resources
drf_yasg

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via ipython
